Refactor capi logic out from NodePool controller
This moves the CAPI related logic into their own file and add test coverage particularly for the reconcile function.
Additional refactor of the business logic itself is left out intentionally for now to contain the scope of the refactor and avoid backward compatibility issues.
