+++ hypershift-operator/controllers/nodepool/manifests.go
@@ -3,11 +3,9 @@ package nodepool
 import (
 	"fmt"
 
-	hyperv1 "github.com/openshift/hypershift/api/hypershift/v1beta1"
 	"github.com/openshift/hypershift/support/util"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	capiv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 )
 
 const (
@@ -19,33 +17,6 @@ const (
 	QualifiedNameMaxLength = 63
 )
 
-func machineDeployment(nodePool *hyperv1.NodePool, controlPlaneNamespace string) *capiv1.MachineDeployment {
-	return &capiv1.MachineDeployment{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      nodePool.GetName(),
-			Namespace: controlPlaneNamespace,
-		},
-	}
-}
-
-func machineSet(nodePool *hyperv1.NodePool, controlPlaneNamespace string) *capiv1.MachineSet {
-	return &capiv1.MachineSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      nodePool.GetName(),
-			Namespace: controlPlaneNamespace,
-		},
-	}
-}
-
-func machineHealthCheck(nodePool *hyperv1.NodePool, controlPlaneNamespace string) *capiv1.MachineHealthCheck {
-	return &capiv1.MachineHealthCheck{
-		TypeMeta: metav1.TypeMeta{},
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      nodePool.GetName(),
-			Namespace: controlPlaneNamespace,
-		},
-	}
-}
 func TunedConfigMap(namespace, name string) *corev1.ConfigMap {
 	return &corev1.ConfigMap{
 		ObjectMeta: metav1.ObjectMeta{
