+++ hypershift-operator/controllers/nodepool/nodepool_controller_test.go
@@ -2,28 +2,22 @@ package nodepool
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"regexp"
 	"strings"
 	"testing"
-	"time"
 
-	"github.com/go-logr/logr"
 	"github.com/google/go-cmp/cmp"
 	. "github.com/onsi/gomega"
 	configv1 "github.com/openshift/api/config/v1"
 	hyperv1 "github.com/openshift/hypershift/api/hypershift/v1beta1"
 	"github.com/openshift/hypershift/api/util/ipnet"
 	"github.com/openshift/hypershift/hypershift-operator/controllers/hostedcluster"
-	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests"
 	ignserver "github.com/openshift/hypershift/ignition-server/controllers"
 	kvinfra "github.com/openshift/hypershift/kubevirtexternalinfra"
 	"github.com/openshift/hypershift/support/api"
 	"github.com/openshift/hypershift/support/releaseinfo"
 	fakereleaseprovider "github.com/openshift/hypershift/support/releaseinfo/fake"
-	"github.com/openshift/hypershift/support/testutil"
-	"github.com/openshift/hypershift/support/upsert"
 	corev1 "k8s.io/api/core/v1"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -31,10 +25,8 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/apimachinery/pkg/util/intstr"
 	"k8s.io/utils/ptr"
-	capiaws "sigs.k8s.io/cluster-api-provider-aws/v2/api/v1beta2"
 	capiv1 "sigs.k8s.io/cluster-api/api/v1beta1"
 	"sigs.k8s.io/controller-runtime/pkg/client"
-	"sigs.k8s.io/controller-runtime/pkg/client/apiutil"
 	"sigs.k8s.io/controller-runtime/pkg/client/fake"
 	"sigs.k8s.io/controller-runtime/pkg/reconcile"
 )
@@ -164,219 +156,6 @@ func TestIsAutoscalingEnabled(t *testing.T) {
 	}
 }
 
-func TestSetMachineDeploymentReplicas(t *testing.T) {
-	testCases := []struct {
-		name                        string
-		nodePool                    *hyperv1.NodePool
-		machineDeployment           *capiv1.MachineDeployment
-		expectReplicas              int32
-		expectAutoscalerAnnotations map[string]string
-	}{
-		{
-			name: "it sets replicas when autoscaling is disabled",
-			nodePool: &hyperv1.NodePool{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: hyperv1.NodePoolSpec{
-					Replicas: ptr.To[int32](5),
-				},
-			},
-			machineDeployment: &capiv1.MachineDeployment{
-				ObjectMeta: metav1.ObjectMeta{
-					CreationTimestamp: metav1.Now(),
-				},
-			},
-			expectReplicas: 5,
-			expectAutoscalerAnnotations: map[string]string{
-				autoscalerMinAnnotation: "0",
-				autoscalerMaxAnnotation: "0",
-			},
-		},
-		{
-			name: "it keeps current replicas and set annotations when autoscaling is enabled",
-			nodePool: &hyperv1.NodePool{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: hyperv1.NodePoolSpec{
-					AutoScaling: &hyperv1.NodePoolAutoScaling{
-						Min: 1,
-						Max: 5,
-					},
-				},
-			},
-			machineDeployment: &capiv1.MachineDeployment{
-				ObjectMeta: metav1.ObjectMeta{
-					CreationTimestamp: metav1.Now(),
-				},
-				Spec: capiv1.MachineDeploymentSpec{
-					Replicas: ptr.To[int32](3),
-				},
-			},
-			expectReplicas: 3,
-			expectAutoscalerAnnotations: map[string]string{
-				autoscalerMinAnnotation: "1",
-				autoscalerMaxAnnotation: "5",
-			},
-		},
-		{
-			name: "it sets current replicas to 1 and set annotations when autoscaling is enabled" +
-				" and the MachineDeployment has not been created yet",
-			nodePool: &hyperv1.NodePool{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: hyperv1.NodePoolSpec{
-					AutoScaling: &hyperv1.NodePoolAutoScaling{
-						Min: 1,
-						Max: 5,
-					},
-				},
-			},
-			machineDeployment: &capiv1.MachineDeployment{},
-			expectReplicas:    1,
-			expectAutoscalerAnnotations: map[string]string{
-				autoscalerMinAnnotation: "1",
-				autoscalerMaxAnnotation: "5",
-			},
-		},
-		{
-			name: "it sets current replicas to 1 and set annotations when autoscaling is enabled" +
-				" and the MachineDeployment has 0 replicas",
-			nodePool: &hyperv1.NodePool{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: hyperv1.NodePoolSpec{
-					AutoScaling: &hyperv1.NodePoolAutoScaling{
-						Min: 1,
-						Max: 5,
-					},
-				},
-			},
-			machineDeployment: &capiv1.MachineDeployment{
-				ObjectMeta: metav1.ObjectMeta{
-					CreationTimestamp: metav1.Now(),
-				},
-				Spec: capiv1.MachineDeploymentSpec{
-					Replicas: ptr.To[int32](0),
-				},
-			},
-			expectReplicas: 1,
-			expectAutoscalerAnnotations: map[string]string{
-				autoscalerMinAnnotation: "1",
-				autoscalerMaxAnnotation: "5",
-			},
-		},
-		{
-			name: "it sets current replicas to 1 and set annotations when autoscaling is enabled" +
-				" and the MachineDeployment has nil replicas",
-			nodePool: &hyperv1.NodePool{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: hyperv1.NodePoolSpec{
-					AutoScaling: &hyperv1.NodePoolAutoScaling{
-						Min: 1,
-						Max: 5,
-					},
-				},
-			},
-			machineDeployment: &capiv1.MachineDeployment{
-				ObjectMeta: metav1.ObjectMeta{
-					CreationTimestamp: metav1.Now(),
-				},
-				Spec: capiv1.MachineDeploymentSpec{
-					Replicas: nil,
-				},
-			},
-			expectReplicas: 1,
-			expectAutoscalerAnnotations: map[string]string{
-				autoscalerMinAnnotation: "1",
-				autoscalerMaxAnnotation: "5",
-			},
-		},
-		{
-			name: "it does not set current replicas but set annotations when autoscaling is enabled" +
-				" and the MachineDeployment has nil replicas",
-			nodePool: &hyperv1.NodePool{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: hyperv1.NodePoolSpec{
-					AutoScaling: &hyperv1.NodePoolAutoScaling{
-						Min: 2,
-						Max: 5,
-					},
-				},
-			},
-			machineDeployment: &capiv1.MachineDeployment{
-				ObjectMeta: metav1.ObjectMeta{
-					CreationTimestamp: metav1.Now(),
-				},
-				Spec: capiv1.MachineDeploymentSpec{
-					Replicas: nil,
-				},
-			},
-			expectReplicas: 2,
-			expectAutoscalerAnnotations: map[string]string{
-				autoscalerMinAnnotation: "2",
-				autoscalerMaxAnnotation: "5",
-			},
-		},
-		{
-			name: "it sets current replicas to autoScaling.min and set annotations when autoscaling is enabled" +
-				" and the MachineDeployment has replicas < autoScaling.min",
-			nodePool: &hyperv1.NodePool{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: hyperv1.NodePoolSpec{
-					AutoScaling: &hyperv1.NodePoolAutoScaling{
-						Min: 2,
-						Max: 5,
-					},
-				},
-			},
-			machineDeployment: &capiv1.MachineDeployment{
-				ObjectMeta: metav1.ObjectMeta{
-					CreationTimestamp: metav1.Now(),
-				},
-				Spec: capiv1.MachineDeploymentSpec{
-					Replicas: ptr.To[int32](1),
-				},
-			},
-			expectReplicas: 2,
-			expectAutoscalerAnnotations: map[string]string{
-				autoscalerMinAnnotation: "2",
-				autoscalerMaxAnnotation: "5",
-			},
-		},
-		{
-			name: "it sets current replicas to autoScaling.max and set annotations when autoscaling is enabled" +
-				" and the MachineDeployment has replicas > autoScaling.max",
-			nodePool: &hyperv1.NodePool{
-				ObjectMeta: metav1.ObjectMeta{},
-				Spec: hyperv1.NodePoolSpec{
-					AutoScaling: &hyperv1.NodePoolAutoScaling{
-						Min: 2,
-						Max: 5,
-					},
-				},
-			},
-			machineDeployment: &capiv1.MachineDeployment{
-				ObjectMeta: metav1.ObjectMeta{
-					CreationTimestamp: metav1.Now(),
-				},
-				Spec: capiv1.MachineDeploymentSpec{
-					Replicas: ptr.To[int32](10),
-				},
-			},
-			expectReplicas: 5,
-			expectAutoscalerAnnotations: map[string]string{
-				autoscalerMinAnnotation: "2",
-				autoscalerMaxAnnotation: "5",
-			},
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			g := NewWithT(t)
-			setMachineDeploymentReplicas(tc.nodePool, tc.machineDeployment)
-			g.Expect(*tc.machineDeployment.Spec.Replicas).To(Equal(tc.expectReplicas))
-			g.Expect(tc.machineDeployment.Annotations).To(Equal(tc.expectAutoscalerAnnotations))
-		})
-	}
-}
-
 func TestValidateManagement(t *testing.T) {
 	intstrPointer1 := intstr.FromInt(1)
 	testCases := []struct {
@@ -494,305 +273,6 @@ func TestValidateManagement(t *testing.T) {
 			g.Expect(err).ShouldNot(HaveOccurred())
 		})
 	}
-}
-
-// It returns a expected machineTemplateSpecJSON
-// and a template and mutateTemplate able to produce an expected target template.
-func RunTestMachineTemplateBuilders(t *testing.T, preCreateMachineTemplate bool) {
-	g := NewWithT(t)
-	c := fake.NewClientBuilder().WithScheme(api.Scheme).WithObjects().Build()
-	r := &NodePoolReconciler{
-		Client:                 c,
-		CreateOrUpdateProvider: upsert.New(false),
-	}
-
-	infraID := "test"
-	ami := "test"
-	hcluster := &hyperv1.HostedCluster{
-		TypeMeta:   metav1.TypeMeta{},
-		ObjectMeta: metav1.ObjectMeta{},
-		Spec: hyperv1.HostedClusterSpec{
-			Release: hyperv1.Release{},
-			InfraID: infraID,
-			Platform: hyperv1.PlatformSpec{
-				Type: hyperv1.AWSPlatform,
-				AWS: &hyperv1.AWSPlatformSpec{
-					ResourceTags: nil,
-				},
-			},
-		},
-		Status: hyperv1.HostedClusterStatus{
-			Platform: &hyperv1.PlatformStatus{
-				AWS: &hyperv1.AWSPlatformStatus{
-					DefaultWorkerSecurityGroupID: "default-sg",
-				},
-			},
-		},
-	}
-	nodePool := &hyperv1.NodePool{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test",
-			Namespace: "test",
-		},
-		Spec: hyperv1.NodePoolSpec{
-			Platform: hyperv1.NodePoolPlatform{
-
-				Type: hyperv1.AWSPlatform,
-				AWS: &hyperv1.AWSNodePoolPlatform{
-					InstanceType:    "",
-					InstanceProfile: "",
-					Subnet:          hyperv1.AWSResourceReference{ID: ptr.To("subnet-xyz")},
-					AMI:             ami,
-					RootVolume: &hyperv1.Volume{
-						Size: 16,
-						Type: "io1",
-						IOPS: 5000,
-					},
-					ResourceTags: nil,
-				},
-			},
-		},
-	}
-
-	if preCreateMachineTemplate {
-		preCreatedMachineTemplate := &capiaws.AWSMachineTemplate{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      nodePool.GetName(),
-				Namespace: manifests.HostedControlPlaneNamespace(hcluster.Namespace, hcluster.Name),
-			},
-			Spec: capiaws.AWSMachineTemplateSpec{
-				Template: capiaws.AWSMachineTemplateResource{
-					Spec: capiaws.AWSMachineSpec{
-						AMI: capiaws.AMIReference{
-							ID: ptr.To(ami),
-						},
-						IAMInstanceProfile:   "test-worker-profile",
-						Subnet:               &capiaws.AWSResourceReference{},
-						UncompressedUserData: ptr.To(true),
-					},
-				},
-			},
-		}
-		err := r.Create(context.Background(), preCreatedMachineTemplate)
-		g.Expect(err).ToNot(HaveOccurred())
-	}
-
-	expectedMachineTemplate := &capiaws.AWSMachineTemplate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        nodePool.GetName(),
-			Namespace:   manifests.HostedControlPlaneNamespace(hcluster.Namespace, hcluster.Name),
-			Annotations: map[string]string{nodePoolAnnotation: client.ObjectKeyFromObject(nodePool).String()},
-		},
-		Spec: capiaws.AWSMachineTemplateSpec{
-			Template: capiaws.AWSMachineTemplateResource{
-				Spec: capiaws.AWSMachineSpec{
-					AMI: capiaws.AMIReference{
-						ID: ptr.To(ami),
-					},
-					IAMInstanceProfile: "test-worker-profile",
-					Subnet: &capiaws.AWSResourceReference{
-						ID: ptr.To("subnet-xyz"),
-					},
-					UncompressedUserData: ptr.To(true),
-					CloudInit: capiaws.CloudInit{
-						InsecureSkipSecretsManager: true,
-						SecureSecretsBackend:       "secrets-manager",
-					},
-					AdditionalTags: capiaws.Tags{
-						awsClusterCloudProviderTagKey(infraID): infraLifecycleOwned,
-					},
-					AdditionalSecurityGroups: []capiaws.AWSResourceReference{
-						{
-							ID: ptr.To("default-sg"),
-						},
-					},
-					RootVolume: &capiaws.Volume{
-						Size: 16,
-						Type: "io1",
-						IOPS: 5000,
-					},
-					InstanceMetadataOptions: &capiaws.InstanceMetadataOptions{
-						HTTPTokens:              capiaws.HTTPTokensStateOptional,
-						HTTPPutResponseHopLimit: 2,
-						HTTPEndpoint:            capiaws.InstanceMetadataEndpointStateEnabled,
-						InstanceMetadataTags:    capiaws.InstanceMetadataEndpointStateDisabled,
-					},
-				},
-			},
-		},
-	}
-	expectedMachineTemplateSpecJSON, err := json.Marshal(expectedMachineTemplate.Spec)
-	g.Expect(err).ToNot(HaveOccurred())
-
-	expectedMachineTemplate.SetName(generateMachineTemplateName(nodePool, expectedMachineTemplateSpecJSON))
-
-	template, mutateTemplate, machineTemplateSpecJSON, err := machineTemplateBuilders(hcluster, nodePool, infraID, ami, "", nil, true)
-	g.Expect(err).ToNot(HaveOccurred())
-	g.Expect(machineTemplateSpecJSON).To(BeIdenticalTo(string(expectedMachineTemplateSpecJSON)))
-
-	// Validate that template and mutateTemplate are able to produce an expected target template.
-	_, err = r.CreateOrUpdate(context.Background(), r.Client, template, func() error {
-		return mutateTemplate(template)
-	})
-	g.Expect(err).ToNot(HaveOccurred())
-
-	gotMachineTemplate := &capiaws.AWSMachineTemplate{}
-	g.Expect(r.Client.Get(context.Background(), client.ObjectKeyFromObject(expectedMachineTemplate), gotMachineTemplate)).To(Succeed())
-	g.Expect(expectedMachineTemplate.Spec).To(BeEquivalentTo(gotMachineTemplate.Spec))
-	g.Expect(expectedMachineTemplate.ObjectMeta.Annotations).To(BeEquivalentTo(gotMachineTemplate.ObjectMeta.Annotations))
-}
-
-func TestMachineTemplateBuilders(t *testing.T) {
-	RunTestMachineTemplateBuilders(t, false)
-}
-
-func TestMachineTemplateBuildersPreexisting(t *testing.T) {
-	RunTestMachineTemplateBuilders(t, true)
-}
-
-func TestCleanupMachineTemplates(t *testing.T) {
-	g := NewWithT(t)
-
-	nodePool := &hyperv1.NodePool{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test",
-			Namespace: "test",
-		},
-		Spec: hyperv1.NodePoolSpec{
-			Platform: hyperv1.NodePoolPlatform{
-				Type: hyperv1.AWSPlatform,
-			},
-		},
-	}
-
-	template1 := &capiaws.AWSMachineTemplate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        "template1",
-			Namespace:   "test",
-			Annotations: map[string]string{nodePoolAnnotation: client.ObjectKeyFromObject(nodePool).String()},
-		},
-		Spec: capiaws.AWSMachineTemplateSpec{},
-	}
-
-	template2 := &capiaws.AWSMachineTemplate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        "template2",
-			Namespace:   "test",
-			Annotations: map[string]string{nodePoolAnnotation: client.ObjectKeyFromObject(nodePool).String()},
-		},
-		Spec: capiaws.AWSMachineTemplateSpec{},
-	}
-
-	gvk, err := apiutil.GVKForObject(template1, api.Scheme)
-	g.Expect(err).ToNot(HaveOccurred())
-	// machine set refrencing template1
-	ms := &capiv1.MachineSet{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        "machineSet",
-			Namespace:   "test",
-			Annotations: map[string]string{nodePoolAnnotation: client.ObjectKeyFromObject(nodePool).String()},
-		},
-		Spec: capiv1.MachineSetSpec{
-			Template: capiv1.MachineTemplateSpec{
-				Spec: capiv1.MachineSpec{
-					InfrastructureRef: corev1.ObjectReference{
-						Kind:       gvk.Kind,
-						APIVersion: gvk.GroupVersion().String(),
-						Name:       template1.Name,
-						Namespace:  template1.Namespace,
-					},
-				},
-			},
-		},
-	}
-
-	c := fake.NewClientBuilder().WithScheme(api.Scheme).WithObjects(nodePool, template1, template2, ms).Build()
-	r := &NodePoolReconciler{
-		Client:                 c,
-		CreateOrUpdateProvider: upsert.New(false),
-	}
-
-	err = r.cleanupMachineTemplates(context.Background(), logr.Discard(), nodePool, "test")
-	g.Expect(err).ToNot(HaveOccurred())
-
-	templates, err := r.listMachineTemplates(nodePool)
-	g.Expect(err).ToNot(HaveOccurred())
-	// check template2 has been deleted
-	g.Expect(len(templates)).To(Equal(1))
-	g.Expect(templates[0].GetName()).To(Equal("template1"))
-}
-
-func TestListMachineTemplatesAWS(t *testing.T) {
-	g := NewWithT(t)
-	capiaws.AddToScheme(api.Scheme)
-	c := fake.NewClientBuilder().WithScheme(api.Scheme).WithObjects().Build()
-	r := &NodePoolReconciler{
-		Client:                 c,
-		CreateOrUpdateProvider: upsert.New(false),
-	}
-	nodePool := &hyperv1.NodePool{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test",
-			Namespace: "test",
-		},
-		Spec: hyperv1.NodePoolSpec{
-			Platform: hyperv1.NodePoolPlatform{
-				Type: hyperv1.AWSPlatform,
-			},
-		},
-	}
-	g.Expect(r.Client.Create(context.Background(), nodePool)).To(BeNil())
-
-	// MachineTemplate with the expected annotation
-	template1 := &capiaws.AWSMachineTemplate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:        "template1",
-			Namespace:   "test",
-			Annotations: map[string]string{nodePoolAnnotation: client.ObjectKeyFromObject(nodePool).String()},
-		},
-		Spec: capiaws.AWSMachineTemplateSpec{},
-	}
-	g.Expect(r.Client.Create(context.Background(), template1)).To(BeNil())
-
-	// MachineTemplate without the expected annoation
-	template2 := &capiaws.AWSMachineTemplate{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "template2",
-			Namespace: "test",
-		},
-		Spec: capiaws.AWSMachineTemplateSpec{},
-	}
-	g.Expect(r.Client.Create(context.Background(), template2)).To(BeNil())
-
-	templates, err := r.listMachineTemplates(nodePool)
-	g.Expect(err).ToNot(HaveOccurred())
-	g.Expect(len(templates)).To(Equal(1))
-	g.Expect(templates[0].GetName()).To(Equal("template1"))
-}
-
-func TestListMachineTemplatesIBMCloud(t *testing.T) {
-	g := NewWithT(t)
-	c := fake.NewClientBuilder().WithScheme(api.Scheme).WithObjects().Build()
-	r := &NodePoolReconciler{
-		Client:                 c,
-		CreateOrUpdateProvider: upsert.New(false),
-	}
-	nodePool := &hyperv1.NodePool{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:      "test",
-			Namespace: "test",
-		},
-		Spec: hyperv1.NodePoolSpec{
-			Platform: hyperv1.NodePoolPlatform{
-				Type: hyperv1.IBMCloudPlatform,
-			},
-		},
-	}
-	g.Expect(r.Client.Create(context.Background(), nodePool)).To(BeNil())
-
-	templates, err := r.listMachineTemplates(nodePool)
-	g.Expect(err).ToNot(HaveOccurred())
-	g.Expect(len(templates)).To(Equal(0))
 }
 
 func TestValidateInfraID(t *testing.T) {
@@ -935,96 +415,6 @@ func TestNodepoolDeletionDoesntRequireHCluster(t *testing.T) {
 	}
 }
 
-func TestInPlaceUpgradeMaxUnavailable(t *testing.T) {
-	intPointer1 := intstr.FromInt(1)
-	intPointer2 := intstr.FromInt(2)
-	strPointer10 := intstr.FromString("10%")
-	strPointer75 := intstr.FromString("75%")
-	testCases := []struct {
-		name     string
-		nodePool *hyperv1.NodePool
-		expect   int
-	}{
-		{
-			name: "defaults to 1 when no maxUnavailable specified",
-			nodePool: &hyperv1.NodePool{
-				Spec: hyperv1.NodePoolSpec{
-					Management: hyperv1.NodePoolManagement{
-						InPlace: &hyperv1.InPlaceUpgrade{},
-					},
-					Replicas: ptr.To[int32](4),
-				},
-			},
-			expect: 1,
-		},
-		{
-			name: "can handle default value of 1",
-			nodePool: &hyperv1.NodePool{
-				Spec: hyperv1.NodePoolSpec{
-					Management: hyperv1.NodePoolManagement{
-						InPlace: &hyperv1.InPlaceUpgrade{
-							MaxUnavailable: &intPointer1,
-						},
-					},
-					Replicas: ptr.To[int32](4),
-				},
-			},
-			expect: 1,
-		},
-		{
-			name: "can handle other values",
-			nodePool: &hyperv1.NodePool{
-				Spec: hyperv1.NodePoolSpec{
-					Management: hyperv1.NodePoolManagement{
-						InPlace: &hyperv1.InPlaceUpgrade{
-							MaxUnavailable: &intPointer2,
-						},
-					},
-					Replicas: ptr.To[int32](4),
-				},
-			},
-			expect: 2,
-		},
-		{
-			name: "can handle percent values",
-			nodePool: &hyperv1.NodePool{
-				Spec: hyperv1.NodePoolSpec{
-					Management: hyperv1.NodePoolManagement{
-						InPlace: &hyperv1.InPlaceUpgrade{
-							MaxUnavailable: &strPointer75,
-						},
-					},
-					Replicas: ptr.To[int32](4),
-				},
-			},
-			expect: 3,
-		},
-		{
-			name: "can handle roundable values",
-			nodePool: &hyperv1.NodePool{
-				Spec: hyperv1.NodePoolSpec{
-					Management: hyperv1.NodePoolManagement{
-						InPlace: &hyperv1.InPlaceUpgrade{
-							MaxUnavailable: &strPointer10,
-						},
-					},
-					Replicas: ptr.To[int32](4),
-				},
-			},
-			expect: 1,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			g := NewWithT(t)
-			maxUnavailable, err := getInPlaceMaxUnavailable(tc.nodePool)
-			g.Expect(err).ToNot(HaveOccurred())
-			g.Expect(maxUnavailable).To(Equal(tc.expect))
-		})
-	}
-}
-
 func TestCreateValidGeneratedPayloadCondition(t *testing.T) {
 	testCases := []struct {
 		name                    string
@@ -1115,53 +505,6 @@ func TestCreateValidGeneratedPayloadCondition(t *testing.T) {
 			got, err := r.createValidGeneratedPayloadCondition(context.Background(), tc.tokenSecret, 1)
 			g.Expect(err).ToNot(HaveOccurred())
 			g.Expect(got).To(BeEquivalentTo(tc.expectedCondition))
-		})
-	}
-}
-
-func TestTaintsToJSON(t *testing.T) {
-	testCases := []struct {
-		name     string
-		taints   []hyperv1.Taint
-		expected string
-	}{
-		{
-			name:     "",
-			taints:   []hyperv1.Taint{},
-			expected: "[]",
-		},
-		{
-			name: "",
-			taints: []hyperv1.Taint{
-				{
-					Key:    "foo",
-					Value:  "bar",
-					Effect: "any",
-				},
-				{
-					Key:    "foo2",
-					Value:  "bar2",
-					Effect: "any",
-				},
-			},
-			expected: `[{"key":"foo","value":"bar","effect":"any"},{"key":"foo2","value":"bar2","effect":"any"}]`,
-		},
-	}
-
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			g := NewWithT(t)
-			taints, err := taintsToJSON(tc.taints)
-			g.Expect(err).ToNot(HaveOccurred())
-			g.Expect(taints).To(BeEquivalentTo(tc.expected))
-
-			// validate decoding.
-			var coreTaints []corev1.Taint
-			err = json.Unmarshal([]byte(taints), &coreTaints)
-			g.Expect(err).ToNot(HaveOccurred())
-			node := &corev1.Node{}
-			node.Spec.Taints = append(node.Spec.Taints, coreTaints...)
-			g.Expect(node.Spec.Taints).To(ContainElements(coreTaints))
 		})
 	}
 }
@@ -2065,170 +1408,6 @@ func TestIsArchAndPlatformSupported(t *testing.T) {
 		t.Run(tc.name, func(t *testing.T) {
 			g := NewWithT(t)
 			g.Expect(isArchAndPlatformSupported(tc.nodePool)).To(Equal(tc.expect))
-		})
-	}
-}
-
-func TestReconcileMachineHealthCheck(t *testing.T) {
-	hostedcluster := func(opts ...func(client.Object)) *hyperv1.HostedCluster {
-		hc := &hyperv1.HostedCluster{ObjectMeta: metav1.ObjectMeta{Namespace: "ns", Name: "cluster"}}
-		for _, o := range opts {
-			o(hc)
-		}
-		return hc
-	}
-
-	nodepool := func(opts ...func(client.Object)) *hyperv1.NodePool {
-		np := &hyperv1.NodePool{ObjectMeta: metav1.ObjectMeta{Namespace: "ns", Name: "nodepool"}}
-		np.Spec.ClusterName = "cluster"
-		for _, o := range opts {
-			o(np)
-		}
-		return np
-	}
-
-	defaultMaxUnhealthy := intstr.Parse("2")
-	healthcheck := func(opts ...func(*capiv1.MachineHealthCheck)) *capiv1.MachineHealthCheck {
-		mhc := &capiv1.MachineHealthCheck{ObjectMeta: metav1.ObjectMeta{Namespace: "ns-cluster", Name: "nodepool"}}
-		resName := generateName("cluster", "cluster", "nodepool")
-		mhc.Spec = capiv1.MachineHealthCheckSpec{
-			ClusterName: "cluster",
-			Selector: metav1.LabelSelector{
-				MatchLabels: map[string]string{
-					resName: resName,
-				},
-			},
-			UnhealthyConditions: []capiv1.UnhealthyCondition{
-				{
-					Type:   corev1.NodeReady,
-					Status: corev1.ConditionFalse,
-					Timeout: metav1.Duration{
-						Duration: time.Duration(8 * time.Minute),
-					},
-				},
-				{
-					Type:   corev1.NodeReady,
-					Status: corev1.ConditionUnknown,
-					Timeout: metav1.Duration{
-						Duration: time.Duration(8 * time.Minute),
-					},
-				},
-			},
-			MaxUnhealthy: &defaultMaxUnhealthy,
-			NodeStartupTimeout: &metav1.Duration{
-				Duration: 20 * time.Minute,
-			},
-		}
-		for _, o := range opts {
-			o(mhc)
-		}
-		return mhc
-	}
-
-	withTimeoutOverride := func(value string) func(client.Object) {
-		return func(o client.Object) {
-			a := o.GetAnnotations()
-			if a == nil {
-				a = map[string]string{}
-			}
-			a[hyperv1.MachineHealthCheckTimeoutAnnotation] = value
-			o.SetAnnotations(a)
-		}
-	}
-
-	withMaxUnhealthyOverride := func(value string) func(client.Object) {
-		return func(o client.Object) {
-			a := o.GetAnnotations()
-			if a == nil {
-				a = map[string]string{}
-			}
-			a[hyperv1.MachineHealthCheckMaxUnhealthyAnnotation] = value
-			o.SetAnnotations(a)
-		}
-	}
-	withMaxUnhealthy := func(value string) func(*capiv1.MachineHealthCheck) {
-		return func(mhc *capiv1.MachineHealthCheck) {
-			maxUnhealthy := intstr.Parse(value)
-			mhc.Spec.MaxUnhealthy = &maxUnhealthy
-		}
-	}
-	withTimeout := func(d time.Duration) func(*capiv1.MachineHealthCheck) {
-		return func(mhc *capiv1.MachineHealthCheck) {
-			for i := range mhc.Spec.UnhealthyConditions {
-				mhc.Spec.UnhealthyConditions[i].Timeout = metav1.Duration{Duration: d}
-			}
-		}
-	}
-
-	tests := []struct {
-		name     string
-		hc       *hyperv1.HostedCluster
-		np       *hyperv1.NodePool
-		expected *capiv1.MachineHealthCheck
-	}{
-		{
-			name:     "defaults",
-			hc:       hostedcluster(),
-			np:       nodepool(),
-			expected: healthcheck(),
-		},
-		{
-			name:     "timeout override in hc",
-			hc:       hostedcluster(withTimeoutOverride("10m")),
-			np:       nodepool(),
-			expected: healthcheck(withTimeout(10 * time.Minute)),
-		},
-		{
-			name:     "timeout override in np",
-			hc:       hostedcluster(),
-			np:       nodepool(withTimeoutOverride("40m")),
-			expected: healthcheck(withTimeout(40 * time.Minute)),
-		},
-		{
-			name:     "timeout override in both, np takes precedence",
-			hc:       hostedcluster(withTimeoutOverride("10m")),
-			np:       nodepool(withTimeoutOverride("40m")),
-			expected: healthcheck(withTimeout(40 * time.Minute)),
-		},
-		{
-			name:     "invalid timeout override, retains default",
-			hc:       hostedcluster(withTimeoutOverride("foo")),
-			np:       nodepool(),
-			expected: healthcheck(),
-		},
-		{
-			name:     "maxunhealthy override in hc",
-			hc:       hostedcluster(withMaxUnhealthyOverride("10%")),
-			np:       nodepool(),
-			expected: healthcheck(withMaxUnhealthy("10%")),
-		},
-		{
-			name:     "maxunhealthy override in np",
-			hc:       hostedcluster(),
-			np:       nodepool(withMaxUnhealthyOverride("5")),
-			expected: healthcheck(withMaxUnhealthy("5")),
-		},
-		{
-			name:     "maxunhealthy override in both, np takes precedence",
-			hc:       hostedcluster(withMaxUnhealthyOverride("10%")),
-			np:       nodepool(withMaxUnhealthyOverride("5")),
-			expected: healthcheck(withMaxUnhealthy("5")),
-		},
-		{
-			name:     "invalid maxunhealthy override value, default is preserved",
-			hc:       hostedcluster(),
-			np:       nodepool(withMaxUnhealthyOverride("foo")),
-			expected: healthcheck(),
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			g := NewWithT(t)
-			r := &NodePoolReconciler{}
-			mhc := &capiv1.MachineHealthCheck{}
-			r.reconcileMachineHealthCheck(context.Background(), mhc, tt.np, tt.hc, "cluster")
-			g.Expect(mhc.Spec).To(testutil.MatchExpected(tt.expected.Spec))
 		})
 	}
 }
